Merge branch 'master' of https://github.com/SupSuper/OpenXcom into oXc_kL

Conflicts:
	Makefile.am
	install/win/installer.nsi
	install/win/installerlang.nsh
	src/Battlescape/AlienBAIState.cpp
	src/Battlescape/ExplosionBState.cpp
	src/Battlescape/NextTurnState.cpp
	src/CMakeLists.txt
	src/Engine/CatFile.cpp
	src/Engine/CatFile.h
	src/Engine/Game.cpp
	src/Engine/Music.cpp
	src/Engine/Music.h
	src/Engine/RNG.cpp
	src/Engine/RNG.h
	src/Geoscape/GeoscapeState.cpp
	src/Menu/OptionsAudioState.cpp
	src/Menu/SaveGameState.cpp
	src/OpenXcom.2010.vcxproj
	src/Resource/XcomResourcePack.cpp
	src/Ruleset/RuleItem.cpp
	src/Ruleset/Ruleset.cpp
	src/Savegame/SavedGame.cpp
	src/Ufopaedia/ArticleStateArmor.cpp

+++ src/Savegame/SavedGame.cpp
@@ -433,7 +433,7 @@ void SavedGame::load(
 	if (doc["rng"]
 		&& !Options::newSeedOnLoad)
 	{
-		RNG::setSeed(doc["rng"].as<int>());
+		RNG::setSeed(doc["rng"].as<uint64_t>());
 	}
 
 	_monthsPassed			= doc["monthsPassed"].as<int>(_monthsPassed);
